requirements.txt
grpcio==1.41.0
billiard==3.6.4.0
blinker==1.7.0
celery==4.4.7
charset-normalizer==3.3.2
connexion==2.14.0
Flask==2.2.3
Flask-Cors==4.0.0
hiredis==2.2.3
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kombu==4.6.11
MarkupSafe==2.1.3
nltk==3.8.1
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
sentry-sdk==1.34.0
SQLAlchemy==1.3.8
swagger-ui-bundle==0.0.9
Werkzeug==2.3.7
gevent==22.10.1
gunicorn==21.2.0
openai==0.28.1
qdrant-client
pandas==2.1.2
scikit-surprise==1.1.3
